feat: player ranking service refactor
